Update contract - General Terms
Bob, I have created the file in markdown, then saved it as HTML and the copied it into this location. I am not sure this is the best way to do this. Note that these are the General Terms of the contract as proposed by Peter. We also have a Software License and an Order Form.